pyproject.toml
[project]
name = "real2sim-eval"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
  # Core runtime
  "torch==2.7.1",
  "torchvision>=0.21.0",
  "torchcodec==0.4.0",
  "triton==3.3.1",

  # General utilities
  "numpy==1.26.4",
  "scipy>=1.15.0",
  "pandas>=2.2.0",
  "tqdm",
  "matplotlib",
  "seaborn",
  "omegaconf",
  "loguru",
  "ninja",
  "networkx==3.5",
  "hydra-core",
  "warp-lang==1.7.0",
  "kornia",
  "opencv-python",

  # 3D / geometry
  "open3d==0.19.0",
  "trimesh",
  "pyvista",
  "pyrender",
  "sapien==2.2.2",
  "plyfile",
  "kornia",

  # JAX stack
  "jax==0.5.3",
  "jax-cuda12-pjrt==0.5.3",
  "jax-cuda12-plugin==0.5.3",
  "flax==0.10.2",
  "optax==0.2.5",
  "equinox==0.13.0",
  "orbax-checkpoint==0.11.13",

  # ML / vision / transformers
  "transformers==4.52.0",
  "diffusers==0.34.0",
  "accelerate==1.10.1",
  "timm==1.0.9",
  "einops",
  "lpips",

  # Visualization / web
  "gradio==5.16.0",
  "dash",
  "plotly",

  # Logging & experiment tracking
  "tensorboard",
  "wandb==0.21.3",
  "ipdb",
]
